requirements.txt
Flask
APScheduler
redis
python-dotenv
pandas
numpy
